+++ projects/pastanaga-angular/src/lib/popup/popover/popover.component.ts
@@ -1,11 +1,11 @@
 import {
-    Component,
-    OnInit,
     ChangeDetectionStrategy,
-    Renderer2,
-    ElementRef,
     ChangeDetectorRef,
+    Component,
+    ElementRef,
     Inject,
+    OnInit,
+    Renderer2,
 } from '@angular/core';
 import { WINDOW } from '@ng-web-apis/common';
 import { PopupComponent } from '../popup.component';
@@ -17,13 +17,10 @@ export const spacer = 8;
 @Component({
     selector: 'pa-popover',
     templateUrl: './popover.component.html',
-    styleUrls: ['./popover.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class PopoverComponent extends PopupComponent implements OnInit {
     popoverHolder?: HTMLElement;
-
-    hasFlexParent = false;
 
     constructor(
         public override popupService: PopupService,
@@ -83,9 +80,7 @@ export class PopoverComponent extends PopupComponent implements OnInit {
     }
 
     private moveToLeft(popup: any, holderRect: DOMRect) {
-        const translateX = this.hasFlexParent
-            ? `calc(-100% - ${holderRect.width + spacer}px)`
-            : `calc(-100% - ${spacer}px)`;
+        const translateX = `calc(-100% - ${spacer}px)`;
         const translateY = `calc(-50% - ${holderRect.height}px/2)`;
         this.style = {
             ...this.style,
@@ -95,7 +90,7 @@ export class PopoverComponent extends PopupComponent implements OnInit {
     }
 
     private moveToRight(popup: any, holderRect: DOMRect) {
-        const translateX = this.hasFlexParent ? `${spacer}px` : `${holderRect.width + spacer}px`;
+        const translateX = `${holderRect.width + spacer}px`;
         const translateY = `calc(-50% - ${holderRect.height}px/2)`;
         this.style = {
             ...this.style,