+++ projects/pastanaga-angular/src/lib/popup/popover/popover.directive.spec.ts
@@ -1,6 +1,6 @@
 import { ExtendedPopupDirective, PopoverDirective } from './popover.directive';
 import { ElementRef } from '@angular/core';
-import { BreakpointObserver, ViewportMode } from '../../breakpoint-observer/breakpoint.observer';
+import { BreakpointObserver, ViewportMode } from '../../breakpoint-observer';
 import { BehaviorSubject } from 'rxjs';
 import { PopoverComponent } from './popover.component';
 import { waitForAsync } from '@angular/core/testing';
@@ -9,98 +9,70 @@ import { PopupComponent } from '../popup.component';
 describe('PopoverDirective', () => {
     let directive: PopoverDirective;
     let popupDirective: ExtendedPopupDirective;
-    let blockParentRef: ElementRef;
-    let flexParentRef: ElementRef;
-    const window: Window = ({
+    let parentRef: ElementRef;
+    const window: Window = {
         getComputedStyle: jest.fn((element) => element.style),
-    } as any) as Window;
+    } as any as Window;
     let popoverComponent: PopoverComponent;
 
     const createDirective: (parentElement: ElementRef, mode?: ViewportMode) => PopoverDirective = (
         parentElement: ElementRef,
         mode = 'desktop',
     ) => {
-        const breakpoint = ({ currentMode: new BehaviorSubject(mode) } as any) as BreakpointObserver;
+        const breakpoint = { currentMode: new BehaviorSubject(mode) } as any as BreakpointObserver;
         return new PopoverDirective(popupDirective, parentElement, breakpoint, window);
     };
 
     beforeEach(() => {
         popupDirective = {} as ExtendedPopupDirective;
-        blockParentRef = {
+        parentRef = {
             nativeElement: {
                 getBoundingClientRect: jest.fn(),
                 parentElement: { style: { display: 'block' } },
-            },
-        };
-        flexParentRef = {
-            nativeElement: {
-                getBoundingClientRect: jest.fn(),
-                parentElement: { style: { display: 'flex' } },
             },
         };
         popoverComponent = {} as PopoverComponent;
     });
 
     describe('paPopover setter', () => {
-        it('should set hasFlexParent on the directive and the component', () => {
-            directive = createDirective(blockParentRef);
-            directive.paPopover = popoverComponent;
-            expect(directive.hasFlexParent).toBe(false);
-            expect(popoverComponent.hasFlexParent).toBe(false);
-
-            directive = createDirective(flexParentRef);
-            directive.paPopover = popoverComponent;
-            expect(directive.hasFlexParent).toBe(true);
-            expect(popoverComponent.hasFlexParent).toBe(true);
-        });
-
         it('should set directive holder element in the component', () => {
-            directive = createDirective(blockParentRef);
+            directive = createDirective(parentRef);
             directive.paPopover = popoverComponent;
-            expect(popoverComponent.popoverHolder).toBe(blockParentRef.nativeElement);
+            expect(popoverComponent.popoverHolder).toBe(parentRef.nativeElement);
         });
 
         it('should set the component in paPopup property of the popup directive', () => {
-            directive = createDirective(blockParentRef);
+            directive = createDirective(parentRef);
             directive.paPopover = popoverComponent;
             expect(directive.popupDirective.paPopup).toBe(popoverComponent);
         });
     });
 
     describe('isVisibleOnHover', () => {
-        it(
-            'should be false on mobile',
-            waitForAsync(() => {
-                directive = createDirective(blockParentRef, 'mobile');
-                directive.isVisibleOnHover.subscribe((hoverEnabled) => expect(hoverEnabled).toBe(false));
-            }),
-        );
+        it('should be false on mobile', waitForAsync(() => {
+            directive = createDirective(parentRef, 'mobile');
+            directive.isVisibleOnHover.subscribe((hoverEnabled) => expect(hoverEnabled).toBe(false));
+        }));
 
-        it(
-            'should be false on tablet',
-            waitForAsync(() => {
-                directive = createDirective(blockParentRef, 'tablet');
-                directive.isVisibleOnHover.subscribe((hoverEnabled) => expect(hoverEnabled).toBe(false));
-            }),
-        );
+        it('should be false on tablet', waitForAsync(() => {
+            directive = createDirective(parentRef, 'tablet');
+            directive.isVisibleOnHover.subscribe((hoverEnabled) => expect(hoverEnabled).toBe(false));
+        }));
 
-        it(
-            'should be true on desktop',
-            waitForAsync(() => {
-                directive = createDirective(blockParentRef, 'desktop');
-                directive.isVisibleOnHover.subscribe((hoverEnabled) => expect(hoverEnabled).toBe(true));
-            }),
-        );
+        it('should be true on desktop', waitForAsync(() => {
+            directive = createDirective(parentRef, 'desktop');
+            directive.isVisibleOnHover.subscribe((hoverEnabled) => expect(hoverEnabled).toBe(true));
+        }));
     });
 
     describe('onClick', () => {
         it('should set popupPosition on popupDirective', () => {
-            directive = createDirective(blockParentRef);
+            directive = createDirective(parentRef);
             const position = { position: 'fixed' };
             // @ts-ignore access private member
             directive.getPosition = jest.fn(() => position);
 
-            directive.onClick(({ stopPropagation: jest.fn(), preventDefault: jest.fn() } as any) as MouseEvent);
+            directive.onClick({ stopPropagation: jest.fn(), preventDefault: jest.fn() } as any as MouseEvent);
             expect(directive.popupDirective.popupPosition).toBe(position);
         });
     });
@@ -110,11 +82,11 @@ describe('PopoverDirective', () => {
         let mockPopup: PopupComponent;
 
         beforeEach(() => {
-            mockPopup = ({ show: jest.fn() } as any) as PopupComponent;
+            mockPopup = { show: jest.fn() } as any as PopupComponent;
         });
 
         it('should do nothing on mobile', () => {
-            directive = createDirective(blockParentRef, 'mobile');
+            directive = createDirective(parentRef, 'mobile');
             directive.popupDirective.paPopup = mockPopup;
             directive.onHover();
             expect(mockPopup.show).not.toHaveBeenCalled();
@@ -122,7 +94,7 @@ describe('PopoverDirective', () => {
 
         describe('on desktop', () => {
             beforeEach(() => {
-                directive = createDirective(blockParentRef, 'desktop');
+                directive = createDirective(parentRef, 'desktop');
                 directive.popupDirective.paPopup = mockPopup;
                 // @ts-ignore access private member
                 directive.getPosition = jest.fn(() => position);
@@ -151,11 +123,11 @@ describe('PopoverDirective', () => {
         let mockPopup: PopupComponent;
 
         beforeEach(() => {
-            mockPopup = ({ close: jest.fn() } as any) as PopupComponent;
+            mockPopup = { close: jest.fn() } as any as PopupComponent;
         });
 
         it('should do nothing on mobile', () => {
-            directive = createDirective(blockParentRef, 'mobile');
+            directive = createDirective(parentRef, 'mobile');
             directive.popupDirective.paPopup = mockPopup;
             directive.onLeave();
             expect(mockPopup.close).not.toHaveBeenCalled();
@@ -163,7 +135,7 @@ describe('PopoverDirective', () => {
 
         describe('on desktop', () => {
             beforeEach(() => {
-                directive = createDirective(blockParentRef, 'desktop');
+                directive = createDirective(parentRef, 'desktop');
                 directive.popupDirective.paPopup = mockPopup;
             });
 
@@ -192,60 +164,38 @@ describe('PopoverDirective', () => {
             height: 150,
             width: 100,
             top: 50,
+            left: 150,
             bottom: 200,
         };
 
-        describe('on flex parent', () => {
-            beforeEach(() => {
-                flexParentRef.nativeElement.getBoundingClientRect.mockReturnValue(parentRect);
-            });
-
-            it('should position the top of the popover at the top of the parent element', () => {
-                directive = createDirective(flexParentRef);
-                directive.hasFlexParent = true;
-                // @ts-ignore access private member
-                const position = directive.getPosition();
-                expect(position.position).toBe('fixed');
-                expect(position.top).toBe(`${parentRect.top}px`);
-            });
-
-            it('on desktop, should translate Y of 8px and translate X 50% minus the half of the parent width', () => {
-                directive = createDirective(flexParentRef);
-                directive.hasFlexParent = true;
-                // @ts-ignore access private member
-                const position = directive.getPosition();
-                expect(position.transform).toBe(`translateX(calc(-50% - ${parentRect.width}px/2)) translateY(8px)`);
-            });
-
-            it('on tablet, should translate Y of 50% plus 16px and translate X 50% minus the half of the parent width', () => {
-                directive = createDirective(flexParentRef, 'tablet');
-                directive.hasFlexParent = true;
-                // @ts-ignore access private member
-                const position = directive.getPosition();
-                expect(position.transform).toBe(
-                    `translateX(calc(-50% - ${parentRect.width}px/2)) translateY(calc(-50% + 16px))`,
-                );
-            });
+        beforeEach(() => {
+            parentRef.nativeElement.getBoundingClientRect.mockReturnValue(parentRect);
+            directive = createDirective(parentRef);
         });
 
-        describe('on block parent', () => {
-            beforeEach(() => {
-                blockParentRef.nativeElement.getBoundingClientRect.mockReturnValue(parentRect);
-                directive = createDirective(blockParentRef);
-            });
+        it('should position the top of popover at the bottom of the parent element', () => {
+            // @ts-ignore access private member
+            const position = directive.getPosition();
+            expect(position.position).toBe('fixed');
+            expect(position.top).toBe(`${parentRect.bottom}px`);
+        });
 
-            it('should position the top of popover at the bottom of the parent element', () => {
-                // @ts-ignore access private member
-                const position = directive.getPosition();
-                expect(position.position).toBe('fixed');
-                expect(position.top).toBe(`${parentRect.bottom}px`);
-            });
+        it('should position the left of the popover on the left of the parent element', () => {
+            // @ts-ignore access private member
+            const position = directive.getPosition();
+            expect(position.left).toBe(`${parentRect.left}px`);
+            expect(position.position).toBe('fixed');
+        });
 
-            it('should translate Y of 8px and translate X of 50% plus the half of the parent width', () => {
-                // @ts-ignore access private member
-                const position = directive.getPosition();
-                expect(position.transform).toBe(`translateX(calc(-50% + ${parentRect.width}px/2)) translateY(8px)`);
-            });
+        it('should translate Y by the offset (8px by default) and translate X of -50% plus the half of the parent width', () => {
+            // @ts-ignore access private member
+            let position = directive.getPosition();
+            expect(position.transform).toBe(`translateX(calc(-50% + ${parentRect.width}px/2)) translateY(8px)`);
+
+            directive.paPopoverOffset = '4px';
+            // @ts-ignore access private member
+            position = directive.getPosition();
+            expect(position.transform).toBe(`translateX(calc(-50% + ${parentRect.width}px/2)) translateY(4px)`);
         });
     });
 });